Consolidate ironcompres loading
