+++ Connectors/Revit/Speckle.Connectors.RevitShared/DependencyInjection/RevitConnectorModule.cs
@@ -31,7 +31,7 @@ public static class ServiceRegistration
   {
     // Pre-load IronCompress's native Zstd lib (net48 only) at startup so BOTH the artefact send and receive paths can
     // (de)compress parquet — receive reads the bundle before the host builder runs. No-op on net8+ (Revit 2025+).
-    RevitZstdNativeLoader.Ensure();
+    ZstdNativeLoader.Ensure();
 
     serviceCollection.AddConnectors();
     serviceCollection.AddDUI<RevitThreadContext, RevitDocumentStore>();
