+++ Connectors/Revit/Speckle.Connectors.RevitShared/Net48.IronCompress.props
@@ -1,7 +1,7 @@
 <Project>
   <!-- net48 only: deploy IronCompress's native nironcompress.dll next to the plugin so Parquet.Net's Zstd codec
        resolves on .NET Framework. On net48 the bare DllImport does no runtimes/ probing, so the native must sit next
-       to the plugin AND be pre-loaded by full path (RevitZstdNativeLoader). net8+ gets the native automatically via
+       to the plugin AND be pre-loaded by full path (Speckle.Connectors.Common's ZstdNativeLoader). net8+ gets the native automatically via
        its RID, so this props is only imported by the net48 connector projects (Revit 2023/2024). -->
   <ItemGroup>
     <PackageReference Include="IronCompress" GeneratePathProperty="true" PrivateAssets="all" />
